+++ framework/base/ErrorHandler.php
@@ -356,24 +356,6 @@ abstract class ErrorHandler extends Component
     }
 
     /**
-     * Converts an exception into a PHP error.
-     *
-     * This method can be used to convert exceptions inside of methods like `__toString()`
-     * to PHP errors because exceptions cannot be thrown inside of them.
-     * @param \Throwable $exception the exception to convert to a PHP error.
-     * @return never
-     *
-     * @deprecated since 2.0.53. Use conditional exception throwing in `__toString()` methods instead.
-     * For PHP < 7.4: use `trigger_error()` directly with `convertExceptionToString()` method.
-     * For PHP >= 7.4: throw the exception directly as `__toString()` supports exceptions.
-     * This method will be removed in 2.2.0.
-     */
-    public static function convertExceptionToError($exception)
-    {
-        trigger_error(static::convertExceptionToString($exception), E_USER_ERROR);
-    }
-
-    /**
      * Converts an exception into a simple string.
      * @param \Throwable $exception the exception being converted
      * @return string the string representation of the exception.