+++ tests/framework/db/mysql/connection/DeadLockTest.php
@@ -263,11 +263,6 @@ class DeadLockTest extends \yiiunit\framework\db\mysql\ConnectionTest
      */
     private function setErrorHandler(): void
     {
-        if (PHP_VERSION_ID < 70000) {
-            set_error_handler(function ($errno, $errstr, $errfile, $errline): never {
-                throw new \ErrorException($errstr, $errno, $errno, $errfile, $errline);
-            });
-        }
     }
 
     /**